Update eloquent-serialization.md
I have an Eloquent model with an attribute with LONGTEXT. I don't need this attribute default but I need it sometimes. 
I Googled for it: How to hide/exclude a column/attribute of an Eloquent model. I've got this documentation as the result so I believed I can save memory by using attribute hiding. After we updated the database and added the GeoJSON data to the database, the site went down due to memory lacking issues. It came up, this happens due to the LONGTEXT column despite that it's hidden. So I think this should be cleared in the docs, attribute hiding is only a visual function, in real the attribute will take the memory.

Feel free correct my English for better understanding :)